feat(server): implements google access token into rest api invocation
- In a previous commit, I obtained an access token to google via
passport--which, under the hood--fetch said token similarly to how I do
in the example/googleapis-nodejs-client branch
- I then used the token in the authorization header to send an http request to the google REST api
OAuth2 endpoints

I am probably not going to use the REST methods in this application
since I think it may be a little "cleaner" to use the nodejs client library. Surely the http endpoints wouldn't change, but if any sort of implementation did, I'd expect their client library to be updated quickly, which would make this app a little more future proof. I also think the code will look less repetitive in terms of the number of http requests we'll end up using to their apis.