refactor+test(server): re-map userTasks
- getUserTasks() gql query now returns the task id, name, and the user
description + isPublic fields (instead of the nested structure returned
from sequelize); note that the graphql schema had to be updated to
accomodate these changes, and the user API (dataSource) MUST return data
to the graphql resolvers that match said schema~~magic!
- move initialization of the models `from server/src/api/gairos` into
`server/src/db`. The test environment was not being synced properly. In
order to get that to occur, this code was moved here to prevent a
circular dependency which would prevent the models from loading in the
test environment
- now using models to seed the database