feat(server): implement gql query getUserTags(userId)
NOTE: I did not form a reducer for this query, as that takes alot of
work; for now, I'll try to stop making these reducers, as it is adding
extra complexity where the client can simply just ignore specific data;
this will push some of the complexity to the client side, but it should
be easier that way. i.e.) if a name refactor occurs on the server side,
code must be touched in one less place overall, than if we did implement
a reducer; I see reducers being useful for 3rd party APIs or if one
wanted to guarantee a specific schema to appear one way (not having so
many nested data fields), and instead, pulling relevant fields up to one
level.

SIDE NOTE: the reducers that I am referring to are custom methods that
shape data from the database to adhere to the graphql schema which
does NOT 100% mirror the sequelize models... as
opposed to simply returning the entire database result to the graphql,
where both the graphql schema and models (the ASSOCIATIONS ESPECIALLY)
are mirrored and letting the graphql client cherry pick what they want
(which they could still do). I believe this is the simpler approach for
now.

- this will fetch all of a users' tags. Each tag will have a property
"userTaskTags" which contains a list of every task assigned to this task
- the relationships are deep enough to get the tag name and the user's
task metadata (now refered to as userTaskInfo)
- refactor 'taskUser' into 'userTaskInfo', as that makes more sense on
the surface level; this refers to the userTasks table, and is used to
fetch each individual users' description of a task