feat+refactor(server): extrapolate root query/mutation to separate files
feat: version query and mutation to
`server\src\api\gairos\__root__`; Set APP_VERSION in .env-example. This
value is returned for these operations

refactor: aforementioned directory will house the root query and mutation as needed for graphql schema (includes custom sca
definitions). This is designed to keep the rest of the models cleaner

NOTE: of course, each model's graphql schema will need to be sure to
extend the graphql query and mutation types