refactor(server): change colors for userTasks to match google color ID
feat: if you do not specify a `eventColorId` when creating/updating a
userTaskHistory, the color defined in the userTaskHistory is used. If
that color is null (which it shouldn't be, since it is defaulted in the
sequelize model, but just in case it was set as null), fall back to the
color defined in the userTask!